refactor(tests): separate post data and firestore post generator
